move reusable widget to own file
